+++ src/command/Command.java
@@ -1,5 +1,8 @@
 package command;
 
+/**
+ * Enum representing the various commands that can be issued in the Giorgo application.
+ */
 public enum Command {
     LIST,
     MARK,
@@ -12,6 +15,13 @@ public enum Command {
     UNKNOWN,
     DATE;
 
+    /**
+     * Converts a string to a corresponding Command enum value.
+     * If the string does not match any command, returns UNKNOWN.
+     *
+     * @param commandString the string representation of the command
+     * @return the corresponding Command enum value, or UNKNOWN if no match is found
+     */
     public static Command fromString(String commandString) {
         for (Command command : Command.values()) {
             if (command.name().equalsIgnoreCase(commandString)) {
