+++ src/command/Parser.java
@@ -2,23 +2,43 @@ package command;
 
 import exception.InvalidInputException;
 import java.time.LocalDate;
-
-import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
+/**
+ * Parses user input into commands and arguments for the Giorgo application.
+ */
 public class Parser {
 
+    /**
+     * Parses the command from the user input.
+     *
+     * @param userInput the input string from the user
+     * @return the corresponding Command enum value
+     */
     public Command parseCommand(String userInput) {
         String[] parts = userInput.split(" ", 2);
         return Command.fromString(parts[0]);
     }
 
+    /**
+     * Parses the argument from the user input.
+     *
+     * @param userInput the input string from the user
+     * @return the argument part of the input string
+     */
     public String parseArgument(String userInput) {
         String[] parts = userInput.split(" ", 2);
         return parts.length > 1 ? parts[1] : "";
     }
 
+    /**
+     * Parses the arguments for a deadline command.
+     *
+     * @param argument the argument string for the deadline command
+     * @return an array containing the description and date/time
+     * @throws InvalidInputException if the format is invalid
+     */
     public String[] parseDeadlineArguments(String argument) throws InvalidInputException {
         String[] parts = argument.split("/by ");
         if (parts.length != 2) {
@@ -27,6 +47,13 @@ public class Parser {
         return parts;
     }
 
+    /**
+     * Parses the arguments for an event command.
+     *
+     * @param argument the argument string for the event command
+     * @return an array containing the description, start, and end times
+     * @throws InvalidInputException if the format is invalid
+     */
     public String[] parseEventArguments(String argument) throws InvalidInputException {
         String[] parts = argument.split("/from |/to ");
         if (parts.length != 3) {
@@ -35,6 +62,13 @@ public class Parser {
         return parts;
     }
 
+    /**
+     * Parses a date string into a LocalDate object.
+     *
+     * @param date the date string in the format d/M/yyyy
+     * @return the corresponding LocalDate object
+     * @throws InvalidInputException if the date format is invalid
+     */
     public LocalDate parseDate(String date) throws InvalidInputException {
         try {
             return LocalDate.parse(date, DateTimeFormatter.ofPattern("d/M/yyyy"));
