--- src/main/giorgo/exception/InvalidInputException.java
@@ -1,8 +1,0 @@
-package exception;
-
-public class InvalidInputException extends Exception {
-    public InvalidInputException(String message) {
-        super(message);
-        System.out.println("ERROR");
-    }
-}
