+++ src/ui/ui.java
@@ -8,15 +8,22 @@ import java.util.ArrayList;
 import java.util.Scanner;
 import task.Event;
 
-
-
+/**
+ * Handles user interactions in the Giorgo application.
+ */
 public class ui {
     private Scanner scanner;
 
+    /**
+     * Constructs a new UI instance.
+     */
     public ui() {
         this.scanner = new Scanner(System.in);
     }
 
+    /**
+     * Displays the welcome message.
+     */
     public void showWelcome() {
         String logo = "Giorgo";
         System.out.println("____________________________________________________________\n" +
@@ -25,22 +32,38 @@ public class ui {
                 "____________________________________________________________\n");
     }
 
+    /**
+     * Displays the goodbye message.
+     */
     public void showGoodbye() {
         System.out.println("____________________________________________________________\n" +
                 " Bye. Hope to see you again soon!\n" +
                 "____________________________________________________________\n");
     }
 
+    /**
+     * Displays a line separator.
+     */
     public void showLine() {
         System.out.println("____________________________________________________________");
     }
 
+    /**
+     * Displays an error message.
+     *
+     * @param message the error message to display
+     */
     public void showError(String message) {
         System.out.println("____________________________________________________________\n" +
                 message + "\n" +
                 "____________________________________________________________\n");
     }
 
+    /**
+     * Displays the list of tasks.
+     *
+     * @param tasks the list of tasks to display
+     */
     public void showTaskList(ArrayList<task> tasks) {
         if (tasks.isEmpty()) {
             System.out.println("____________________________________________________________\n" +
@@ -56,6 +79,12 @@ public class ui {
         }
     }
 
+    /**
+     * Displays the tasks on a specified date.
+     *
+     * @param date the date to filter tasks by
+     * @param tasks the list of tasks to filter
+     */
     public static void listTasksOnDate(String date, ArrayList<task> tasks) {
         LocalDate specifiedDate = LocalDate.parse(date, DateTimeFormatter.ofPattern("d/M/yyyy"));
         System.out.println("____________________________________________________________\n" +
@@ -71,6 +100,12 @@ public class ui {
         System.out.println("____________________________________________________________\n");
     }
 
+    /**
+     * Displays a message indicating a task has been added.
+     *
+     * @param task the task that was added
+     * @param size the new size of the task list
+     */
     public void showTaskAdded(task task, int size) {
         System.out.println("____________________________________________________________\n" +
                 " Got it. I've added this task:\n" +
@@ -79,6 +114,12 @@ public class ui {
                 "____________________________________________________________\n");
     }
 
+    /**
+     * Displays a message indicating a task has been deleted.
+     *
+     * @param task the task that was deleted
+     * @param size the new size of the task list
+     */
     public void showTaskDeleted(task task, int size) {
         System.out.println("____________________________________________________________\n" +
                 " Noted. I've removed this task:\n" +
@@ -87,6 +128,11 @@ public class ui {
                 "____________________________________________________________\n");
     }
 
+    /**
+     * Displays a message indicating a task has been marked as done.
+     *
+     * @param task the task that was marked as done
+     */
     public void showTaskMarked(task task) {
         System.out.println("____________________________________________________________\n" +
                 " Nice! I've marked this task as done:\n" +
@@ -94,6 +140,11 @@ public class ui {
                 "____________________________________________________________\n");
     }
 
+    /**
+     * Displays a message indicating a task has been marked as not done.
+     *
+     * @param task the task that was marked as not done
+     */
     public void showTaskUnmarked(task task) {
         System.out.println("____________________________________________________________\n" +
                 " OK, I've marked this task as not done yet:\n" +
@@ -101,6 +152,11 @@ public class ui {
                 "____________________________________________________________\n");
     }
 
+    /**
+     * Reads a command from the user.
+     *
+     * @return the command entered by the user
+     */
     public String readCommand() {
         return scanner.nextLine().trim();
     }
